This allows you run the FastBoot server in a mode that extracts API calls and embeds them into the index.html for the initial request.
__Do NOT use this in production in it's current state__

This branch is just an reference on how we could ship Resource Discovery Mode to Fastboot. It currently does not [chunk encode](https://en.wikipedia.org/wiki/Chunked_transfer_encoding) the response but rather collects the API calls and then places them in a script tag in the index.html.

For this to work completely you need to create a client library that is guranteed to be there before Ember starts issuing requests to create Request and Response table. This table is meant to act like a network table or request proxy layer. It's purpose is to match up issued requests from the app to a streamed response.

In hindsight the a ServiceWorker is actually the correct tool for the client library as it is a true network middleman.